+++ mc2/controllers/base/managers/infrastructure.py
@@ -24,7 +24,9 @@ class GeneralInfrastructureManager(object):
         return requests.get('%s/v2/apps/%s' % (
             settings.MESOS_MARATHON_HOST,
             app_id,
-        ), headers=self.headers).json()['app']
+        ),
+            headers=self.headers,
+            timeout=settings.DEFAULT_REQUEST_TIMEOUT).json()['app']
 
     def get_marathon_app_tasks(self, app_id):
         """
@@ -33,10 +35,13 @@ class GeneralInfrastructureManager(object):
         :param app_id str: The application id
         :returns: list
         """
-        return requests.get('%s/v2/apps/%s/tasks' % (
+        app_info = requests.get('%s/v2/apps/%s/tasks' % (
             settings.MESOS_MARATHON_HOST,
             app_id,
-        ), headers=self.headers).json()['tasks']
+        ),
+            headers=self.headers,
+            timeout=settings.DEFAULT_REQUEST_TIMEOUT).json()
+        return app_info.get('tasks', [])
 
     def get_marathon_info(self):
         """
@@ -46,7 +51,8 @@ class GeneralInfrastructureManager(object):
         """
         return requests.get(
             '%s/v2/info' % (settings.MESOS_MARATHON_HOST,),
-            headers=self.headers).json()
+            headers=self.headers,
+            timeout=settings.DEFAULT_REQUEST_TIMEOUT).json()
 
     def get_worker_info(self, hostname):
         """
@@ -56,12 +62,13 @@ class GeneralInfrastructureManager(object):
         """
         return requests.get(
             'http://%s:%s/state.json' % (hostname, settings.MESOS_HTTP_PORT),
-            headers=self.headers
+            headers=self.headers,
+            timeout=settings.DEFAULT_REQUEST_TIMEOUT
         ).json()
 
     def get_app_log_info(self, app_id):
         """
-        Returns a list of task info dicts for logdriver for the given app_id
+        Returns a list of task info dicts for the given app_id
 
         :param app_id str: The application id
         :returns: list
@@ -79,8 +86,8 @@ class GeneralInfrastructureManager(object):
     def get_task_log_info(self, app_id, task_id, task_host,
                           marathon_info=None):
         """
-        Returns a dictionary with the task_host and task_dir for logdriver
-        logs for the given task_id
+        Returns a dictionary with the task_host and task_dir
+        for the given task_id
 
         :param app_id str: the application id
         :param task_id str: the task id
@@ -90,17 +97,20 @@ class GeneralInfrastructureManager(object):
         :returns: dict
         """
         marathon_info = marathon_info or self.get_marathon_info()
+        worker_info = self.get_worker_info(task_host)
         framework_id = marathon_info['frameworkId']
-        follower_id = self.get_worker_info(task_host)['id']
+        [framework_executor] = [framework
+                                for framework in worker_info['frameworks']
+                                if framework['id'] == framework_id]
+
+        [task_info] = [task
+                       for task in framework_executor['executors']
+                       if task["id"] == task_id]
+
         return {
             'task_id': task_id,
             'task_host': task_host,
-            'task_dir': (
-                "%(follower_id)s/frameworks/%(framework_id)s/executors"
-                "/%(task_id)s/runs/latest") % {
-                    'follower_id': follower_id,
-                    'framework_id': framework_id,
-                    'task_id': task_id}
+            'task_dir': task_info["directory"]
         }