+++ mc2/controllers/docker/forms.py
@@ -37,7 +37,8 @@ class DockerControllerForm(ControllerForm):
             'name', 'marathon_cpus', 'marathon_mem', 'marathon_instances',
             'marathon_cmd', 'docker_image', 'marathon_health_check_path',
             'port', 'domain_urls', 'volume_needed', 'volume_path',
-            'webhook_token', 'description')
+            'webhook_token', 'description', 'organization',
+            'postgres_db_needed')
 
 
 class DockerControllerFormHelper(ControllerFormHelper):
@@ -47,4 +48,4 @@ class DockerControllerFormHelper(ControllerFormHelper):
         super(DockerControllerFormHelper, self).__init__(
             data, files, instance, prefix, initial)
         self.controller_form = DockerControllerForm(
-            data, files, instance=instance)
+            data, files, instance=instance, initial=initial)