+++ mc2/controllers/docker/tests/test_docker_controller.py
@@ -1,4 +1,5 @@
 import json
+import string
 
 import pytest
 import responses
@@ -13,7 +14,7 @@ from hypothesis.strategies import text, random_module, lists, just
 
 from mc2.controllers.base.models import EnvVariable, MarathonLabel
 from mc2.controllers.base.tests.base import ControllerBaseTestCase
-from mc2.controllers.docker.models import DockerController
+from mc2.controllers.docker.models import DockerController, traefik_domains
 from mc2.organizations.models import Organization
 
 
@@ -45,11 +46,17 @@ def docker_controller(with_envvars=True, with_labels=True, **kw):
     # so we remove them from the generated value.
     # TODO: Build a proper SlugField strategy.
     # TODO: Figure out why the field validation isn't being applied.
-    slug = text().map(lambda t: "".join(t.replace(":", "").split()))
+    # Slugs must be domain-name friendly - used in the "generic" domain
+    slug = text(string.ascii_letters + string.digits + '-')
     kw.setdefault("slug", slug)
 
     kw.setdefault("owner", models(User, is_active=just(True)))
     kw.setdefault("organization", models(Organization, slug=slug))
+
+    # Prevent Hypothesis from generating domains with invalid characters
+    domain_urls = text(string.ascii_letters + string.digits + '-.')
+    kw.setdefault("domain_urls", domain_urls)
+
     # The model generator sees `controller_ptr` (from the PolymorphicModel
     # magic) as a mandatory field and objects if we don't provide a value for
     # it.
@@ -81,6 +88,7 @@ def check_and_clear_appdata(appdata, controller):
     check_and_remove_health(appdata, controller)
     check_and_remove_env(appdata, controller)
     check_and_remove_labels(appdata, controller)
+    check_and_remove_backoff_params(appdata)
     assert appdata == {}
 
 
@@ -113,7 +121,7 @@ def check_and_remove_health(appdata, controller):
     """
     Assert that the health check data is correct and remove it.
     """
-    if controller.marathon_health_check_path:
+    if controller.marathon_health_check_path and controller.port:
         assert appdata.pop("ports") == [0]
         assert appdata.pop("healthChecks") == [{
             "gracePeriodSeconds": 60,
@@ -126,6 +134,11 @@ def check_and_remove_health(appdata, controller):
         }]
     assert "ports" not in appdata
     assert "healthChecks" not in appdata
+
+
+def check_and_remove_backoff_params(appdata):
+    appdata.pop('backoffFactor')
+    appdata.pop('backoffSeconds')
 
 
 def check_and_remove_env(appdata, controller):
@@ -149,6 +162,14 @@ def check_and_remove_labels(appdata, controller):
     domains = [u".".join([controller.app_id, settings.HUB_DOMAIN])]
     domains.extend(controller.domain_urls.split())
     assert sorted(labels.pop("domain").split()) == sorted(domains)
+    assert sorted(labels.pop("HAPROXY_0_VHOST").split()) == sorted(domains)
+    assert labels.pop("HAPROXY_GROUP") == "external"
+
+    traefik_domains = labels.pop("traefik.frontend.rule")
+    traefik_domains = traefik_domains.split(":", 2)[-1].split(",")
+    traefik_domains = [d.strip() for d in traefik_domains]
+    assert sorted(traefik_domains) == sorted(domains)
+
     # We may have duplicate keys in here, but hopefully the database always
     # return the objects in the same order.
     lvs = {lv.name: lv.value for lv in controller.label_variables.all()}
@@ -162,6 +183,34 @@ def check_and_remove_optional(appdata, field, value):
     if value:
         assert appdata.pop(field) == value
     assert field not in appdata
+
+
+def test_traefik_domains_single():
+    """
+    When a domains string with a single domain is passed to traefik_domains,
+    it returns a Host frontend rule.
+    """
+    domains = 'abc.com'
+    assert traefik_domains(domains) == 'Host: abc.com'
+
+
+def test_traefik_domains_multiple():
+    """
+    When a domains string with multiple domains is passed to traefik_domains,
+    it returns multiple Host frontend rules joined by ';'.
+    """
+    domains = 'abc.com def.co.za   ghi.co.ng'
+    assert (traefik_domains(domains) ==
+            'Host: abc.com, def.co.za, ghi.co.ng')
+
+
+def test_traefik_domains_none():
+    """
+    When a domains string with no domains is passed to traefik_domains, it
+    returns an empty string.
+    """
+    domains = '  '
+    assert traefik_domains(domains) == ''
 
 
 @pytest.mark.django_db
@@ -268,16 +317,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
 
         custom_urls = "testing.com url.com"
         controller.domain_urls += custom_urls
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -293,16 +347,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         controller.port = 1234
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App"
             },
             "container": {
@@ -319,16 +378,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         controller.marathon_health_check_path = '/health/path/'
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -356,16 +420,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         controller.volume_path = "/deploy/media/"
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -399,16 +468,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         controller.volume_path = ""
         controller.save()
 
+        domain_label = "{}.{} {}".format(
+            controller.app_id, settings.HUB_DOMAIN, custom_urls)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{} {}".format(controller.app_id,
-                                            settings.HUB_DOMAIN,
-                                            custom_urls),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -450,16 +524,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         )
         self.mk_env_variable(controller)
 
+        domain_label = "{}.{}".format(controller.app_id, settings.HUB_DOMAIN)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "env": {"TEST_KEY": "a test value"},
             "labels": {
-                "domain": "{}.{}".format(controller.app_id,
-                                         settings.HUB_DOMAIN),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -482,16 +561,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
         self.mk_env_variable(controller)
         self.mk_labels_variable(controller)
 
+        domain_label = "{}.{}".format(controller.app_id, settings.HUB_DOMAIN)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
             "cmd": "ping",
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "env": {"TEST_KEY": "a test value"},
             "labels": {
-                "domain": "{}.{}".format(controller.app_id,
-                                         settings.HUB_DOMAIN),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
                 "TEST_LABELS_NAME": 'a test label value'
             },
@@ -534,14 +618,19 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             docker_image='docker/image',
         )
 
+        domain_label = "{}.{}".format(controller.app_id, settings.HUB_DOMAIN)
         self.assertEquals(controller.get_marathon_app_data(), {
             "id": controller.app_id,
             "cpus": 0.1,
             "mem": 128.0,
             "instances": 1,
+            "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+            "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
             "labels": {
-                "domain": "{}.{}".format(controller.app_id,
-                                         settings.HUB_DOMAIN),
+                "domain": domain_label,
+                "HAPROXY_GROUP": "external",
+                "HAPROXY_0_VHOST": domain_label,
+                "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
             },
             "container": {
@@ -571,16 +660,21 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             MESOS_DEFAULT_GRACE_PERIOD_SECONDS='600',
             MESOS_DEFAULT_INTERVAL_SECONDS='100',
                 MESOS_DEFAULT_TIMEOUT_SECONDS='200'):
+            domain_label = "{}.{} {}".format(
+                controller.app_id, settings.HUB_DOMAIN, custom_urls)
             self.assertEquals(controller.get_marathon_app_data(), {
                 "id": controller.app_id,
                 "cpus": 0.1,
                 "mem": 128.0,
                 "instances": 1,
                 "cmd": "ping",
+                "backoffFactor": settings.MESOS_DEFAULT_BACKOFF_FACTOR,
+                "backoffSeconds": settings.MESOS_DEFAULT_BACKOFF_SECONDS,
                 "labels": {
-                    "domain": "{}.{} {}".format(controller.app_id,
-                                                settings.HUB_DOMAIN,
-                                                custom_urls),
+                    "domain": domain_label,
+                    "HAPROXY_GROUP": "external",
+                    "HAPROXY_0_VHOST": domain_label,
+                    "traefik.frontend.rule": traefik_domains(domain_label),
                     "name": "Test App",
                 },
                 "container": {